renderervk: C++ compatibility for tr_bsp

+++ code/renderervk/tr_bsp.c
@@ -51,7 +51,7 @@ static void HSVtoRGB( float h, float s, float v, float rgb[3] )
 
 	h *= 5;
 
-	i = floor( h );
+	i = floor( (double)(h) );
 	f = h - i;
 
 	p = v * ( 1 - s );
@@ -146,7 +146,7 @@ void R_ColorShiftLightingBytes( const byte in[4], byte out[4], qboolean hasAlpha
 		out[1] = luma;
 		out[2] = luma;
 	} else if( r_mapGreyScale->value ) {
-		const float scale = fabs( r_mapGreyScale->value );
+		const float scale = fabs( (double)(r_mapGreyScale->value) );
 		const float luma = LUMA( r, g, b );
 		out[0] = LERP( r, luma, scale );
 		out[1] = LERP( g, luma, scale );
@@ -951,9 +951,9 @@ static qboolean R_MergedWidthPoints( const srfGridMesh_t *grid, int offset ) {
 
 	for (i = 1; i < grid->width-1; i++) {
 		for (j = i + 1; j < grid->width-1; j++) {
-			if ( fabs(grid->verts[i + offset].xyz[0] - grid->verts[j + offset].xyz[0]) > .1) continue;
-			if ( fabs(grid->verts[i + offset].xyz[1] - grid->verts[j + offset].xyz[1]) > .1) continue;
-			if ( fabs(grid->verts[i + offset].xyz[2] - grid->verts[j + offset].xyz[2]) > .1) continue;
+			if ( fabs((double)(grid->verts[i + offset].xyz[0] - grid->verts[j + offset].xyz[0])) > .1) continue;
+			if ( fabs((double)(grid->verts[i + offset].xyz[1] - grid->verts[j + offset].xyz[1])) > .1) continue;
+			if ( fabs((double)(grid->verts[i + offset].xyz[2] - grid->verts[j + offset].xyz[2])) > .1) continue;
 			return qtrue;
 		}
 	}
@@ -973,9 +973,9 @@ static qboolean R_MergedHeightPoints( const srfGridMesh_t *grid, int offset ) {
 
 	for (i = 1; i < grid->height-1; i++) {
 		for (j = i + 1; j < grid->height-1; j++) {
-			if ( fabs(grid->verts[grid->width * i + offset].xyz[0] - grid->verts[grid->width * j + offset].xyz[0]) > .1) continue;
-			if ( fabs(grid->verts[grid->width * i + offset].xyz[1] - grid->verts[grid->width * j + offset].xyz[1]) > .1) continue;
-			if ( fabs(grid->verts[grid->width * i + offset].xyz[2] - grid->verts[grid->width * j + offset].xyz[2]) > .1) continue;
+			if ( fabs((double)(grid->verts[grid->width * i + offset].xyz[0] - grid->verts[grid->width * j + offset].xyz[0])) > .1) continue;
+			if ( fabs((double)(grid->verts[grid->width * i + offset].xyz[1] - grid->verts[grid->width * j + offset].xyz[1])) > .1) continue;
+			if ( fabs((double)(grid->verts[grid->width * i + offset].xyz[2] - grid->verts[grid->width * j + offset].xyz[2])) > .1) continue;
 			return qtrue;
 		}
 	}
@@ -1024,9 +1024,9 @@ static void R_FixSharedVertexLodError_r( int start, srfGridMesh_t *grid1 ) {
 					if (R_MergedWidthPoints(grid2, offset2)) continue;
 					for ( l = 1; l < grid2->width-1; l++) {
 					//
-						if ( fabs(grid1->verts[k + offset1].xyz[0] - grid2->verts[l + offset2].xyz[0]) > .1) continue;
-						if ( fabs(grid1->verts[k + offset1].xyz[1] - grid2->verts[l + offset2].xyz[1]) > .1) continue;
-						if ( fabs(grid1->verts[k + offset1].xyz[2] - grid2->verts[l + offset2].xyz[2]) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[0] - grid2->verts[l + offset2].xyz[0])) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[1] - grid2->verts[l + offset2].xyz[1])) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[2] - grid2->verts[l + offset2].xyz[2])) > .1) continue;
 						// ok the points are equal and should have the same lod error
 						grid2->widthLodError[l] = grid1->widthLodError[k];
 						touch = qtrue;
@@ -1039,9 +1039,9 @@ static void R_FixSharedVertexLodError_r( int start, srfGridMesh_t *grid1 ) {
 					if (R_MergedHeightPoints(grid2, offset2)) continue;
 					for ( l = 1; l < grid2->height-1; l++) {
 					//
-						if ( fabs(grid1->verts[k + offset1].xyz[0] - grid2->verts[grid2->width * l + offset2].xyz[0]) > .1) continue;
-						if ( fabs(grid1->verts[k + offset1].xyz[1] - grid2->verts[grid2->width * l + offset2].xyz[1]) > .1) continue;
-						if ( fabs(grid1->verts[k + offset1].xyz[2] - grid2->verts[grid2->width * l + offset2].xyz[2]) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[0] - grid2->verts[grid2->width * l + offset2].xyz[0])) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[1] - grid2->verts[grid2->width * l + offset2].xyz[1])) > .1) continue;
+						if ( fabs((double)(grid1->verts[k + offset1].xyz[2] - grid2->verts[grid2->width * l + offset2].xyz[2])) > .1) continue;
 						// ok the points are equal and should have the same lod error
 						grid2->heightLodError[l] = grid1->widthLodError[k];
 						touch = qtrue;
@@ -1062,9 +1062,9 @@ static void R_FixSharedVertexLodError_r( int start, srfGridMesh_t *grid1 ) {
 					if (R_MergedWidthPoints(grid2, offset2)) continue;
 					for ( l = 1; l < grid2->width-1; l++) {
 					//
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[0] - grid2->verts[l + offset2].xyz[0]) > .1) continue;
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[1] - grid2->verts[l + offset2].xyz[1]) > .1) continue;
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[2] - grid2->verts[l + offset2].xyz[2]) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[0] - grid2->verts[l + offset2].xyz[0])) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[1] - grid2->verts[l + offset2].xyz[1])) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[2] - grid2->verts[l + offset2].xyz[2])) > .1) continue;
 						// ok the points are equal and should have the same lod error
 						grid2->widthLodError[l] = grid1->heightLodError[k];
 						touch = qtrue;
@@ -1077,9 +1077,9 @@ static void R_FixSharedVertexLodError_r( int start, srfGridMesh_t *grid1 ) {
 					if (R_MergedHeightPoints(grid2, offset2)) continue;
 					for ( l = 1; l < grid2->height-1; l++) {
 					//
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[0] - grid2->verts[grid2->width * l + offset2].xyz[0]) > .1) continue;
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[1] - grid2->verts[grid2->width * l + offset2].xyz[1]) > .1) continue;
-						if ( fabs(grid1->verts[grid1->width * k + offset1].xyz[2] - grid2->verts[grid2->width * l + offset2].xyz[2]) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[0] - grid2->verts[grid2->width * l + offset2].xyz[0])) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[1] - grid2->verts[grid2->width * l + offset2].xyz[1])) > .1) continue;
+						if ( fabs((double)(grid1->verts[grid1->width * k + offset1].xyz[2] - grid2->verts[grid2->width * l + offset2].xyz[2])) > .1) continue;
 						// ok the points are equal and should have the same lod error
 						grid2->heightLodError[l] = grid1->heightLodError[k];
 						touch = qtrue;
@@ -1156,27 +1156,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[k + offset1].xyz;
 					v2 = grid2->verts[l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[k + 2 + offset1].xyz;
 					v2 = grid2->verts[l + 1 + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[l + offset2].xyz;
 					v2 = grid2->verts[l + 1 + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1200,27 +1200,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 					//
 					v1 = grid1->verts[k + offset1].xyz;
 					v2 = grid2->verts[grid2->width * l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[k + 2 + offset1].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[grid2->width * l + offset2].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1253,27 +1253,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[grid1->width * k + offset1].xyz;
 					v2 = grid2->verts[l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[grid1->width * (k + 2) + offset1].xyz;
 					v2 = grid2->verts[l + 1 + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[l + offset2].xyz;
 					v2 = grid2->verts[(l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1297,27 +1297,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[grid1->width * k + offset1].xyz;
 					v2 = grid2->verts[grid2->width * l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[grid1->width * (k + 2) + offset1].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[grid2->width * l + offset2].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1351,27 +1351,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[k + offset1].xyz;
 					v2 = grid2->verts[l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[k - 2 + offset1].xyz;
 					v2 = grid2->verts[l + 1 + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[l + offset2].xyz;
 					v2 = grid2->verts[(l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1395,27 +1395,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[k + offset1].xyz;
 					v2 = grid2->verts[grid2->width * l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[k - 2 + offset1].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[grid2->width * l + offset2].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1450,27 +1450,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[grid1->width * k + offset1].xyz;
 					v2 = grid2->verts[l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[grid1->width * (k - 2) + offset1].xyz;
 					v2 = grid2->verts[l + 1 + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[l + offset2].xyz;
 					v2 = grid2->verts[(l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -1494,27 +1494,27 @@ static int R_StitchPatches( int grid1num, int grid2num ) {
 				//
 					v1 = grid1->verts[grid1->width * k + offset1].xyz;
 					v2 = grid2->verts[grid2->width * l + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 
 					v1 = grid1->verts[grid1->width * (k - 2) + offset1].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) > .1)
+					if ( fabs((double)(v1[0] - v2[0])) > .1)
 						continue;
-					if ( fabs(v1[1] - v2[1]) > .1)
+					if ( fabs((double)(v1[1] - v2[1])) > .1)
 						continue;
-					if ( fabs(v1[2] - v2[2]) > .1)
+					if ( fabs((double)(v1[2] - v2[2])) > .1)
 						continue;
 					//
 					v1 = grid2->verts[grid2->width * l + offset2].xyz;
 					v2 = grid2->verts[grid2->width * (l + 1) + offset2].xyz;
-					if ( fabs(v1[0] - v2[0]) < .01 &&
-							fabs(v1[1] - v2[1]) < .01 &&
-							fabs(v1[2] - v2[2]) < .01)
+					if ( fabs((double)(v1[0] - v2[0])) < .01 &&
+							fabs((double)(v1[1] - v2[1])) < .01 &&
+							fabs((double)(v1[2] - v2[2])) < .01)
 						continue;
 					//
 					//ri.Printf( PRINT_ALL, "found highest LoD crack between two patches\n" );
@@ -2221,8 +2221,8 @@ static void R_LoadLightGrid( const lump_t *l ) {
 	wMaxs = w->bmodels[0].bounds[1];
 
 	for ( i = 0 ; i < 3 ; i++ ) {
-		w->lightGridOrigin[i] = w->lightGridSize[i] * ceil( wMins[i] / w->lightGridSize[i] );
-		maxs[i] = w->lightGridSize[i] * floor( wMaxs[i] / w->lightGridSize[i] );
+		w->lightGridOrigin[i] = w->lightGridSize[i] * ceil( (double)(wMins[i] / w->lightGridSize[i]) );
+		maxs[i] = w->lightGridSize[i] * floor( (double)(wMaxs[i] / w->lightGridSize[i]) );
 		bounds[i] = (maxs[i] - w->lightGridOrigin[i])/w->lightGridSize[i] + 1;
 	}
